refactor(core): Consolidate protocols and remove dead code
- Move IConstraint from constraints/ to base/protocols.py
- Fix decorator placement (stream_validator, udf_validator) at module level
- Remove empty/unused files: base/constraint.py, constraints/, logic/validators.py
- Clean up unused import in models/validation.py
- Update imports to use new protocol locations

This completes the core restructuring with all protocols now centralized in base/protocols.py and eliminates orphaned constraint modules.
